New sequencing run: 20240709
Sequencing run is a re-sequencing of libraries from 2024-06-20 and
2024-06-25. This commit only processes reads from the 2024-07-09
sequencing run from the 2024-06-20 library. Because of the different
read lengths, the reads were analyzed in a new notebook, 20240709.ipynb.
The notebook for the previous reads, 20240620.ipynb, was rerun with an
updated set of packages (no actual results should have changed).